# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts book",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This backend documentation is describing how you can manage your contacts"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://contacts-with-crud.onrender.com"
    }
  ],
  "tags": [
    {
      "name": "Contacts",
      "description": "Contacts operations."
    },
    {
      "name": "Auth",
      "description": "Authorisation."
    }
  ],
  "paths": {
    "/contacts/{contactId}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contacts by it's ID",
        "operationId": "getContact",
        "description": "You can add ID to find a specific contact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66d9bd6a3b0b0e7e161e1330"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found a contact with it`s ID",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id 66d9bd6a3b0b0e7e161e1330!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update the contact",
        "operationId": "updateContact",
        "description": "Here you change details of existing contact",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66d9bd6a3b0b0e7e161e1330"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber",
                  "contactType"
                ],
                "properties": {
                  "name": {
                    "description": "Contact name",
                    "type": "string",
                    "example": "Olivia Prijma"
                  },
                  "phoneNumber": {
                    "description": "Phone number",
                    "type": "string",
                    "example": "+420774839319"
                  },
                  "email": {
                    "description": "Email",
                    "type": "string",
                    "example": "example@mail.com"
                  },
                  "isFavourite": {
                    "description": "Is the contact favourite",
                    "type": "boolean",
                    "example": "true"
                  },
                  "contactType": {
                    "description": "Is the contact favourite",
                    "type": "boolean",
                    "example": "true/false"
                  },
                  "photo": {
                    "description": "Avatar or contact picture",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The successfull",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully patched a contact"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Authorisation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "NotFound"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "message"
                      ],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Action is not allowed"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete a contact by ID",
        "operationId": "deleteContact",
        "description": "You can delete a specific contact by it's ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66d9bd6a3b0b0e7e161e1330"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contact with id 66d9bd6a3b0b0e7e161e1330 deleted",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 204
                    },
                    "data": {}
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get all contacts",
        "operationId": "getAllContacts",
        "description": "Get all contacts of the logged-in user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found contacts!",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id 66d9bd6a3b0b0e7e161e1330!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "perPage": {
                      "type": "integer",
                      "example": 10
                    },
                    "totalItems": {
                      "type": "integer",
                      "example": 56
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 6
                    },
                    "hasPreviousPage": {
                      "type": "boolean",
                      "example": "true/false"
                    },
                    "hasNextPage": {
                      "type": "boolean",
                      "example": "true/false"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorisation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "UnauthorizedError"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "message"
                      ],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access token expired"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Create contact",
        "operationId": "createContact",
        "description": "Here you can add new contacts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phoneNumber",
                  "contactType"
                ],
                "properties": {
                  "name": {
                    "description": "Contact name",
                    "type": "string",
                    "example": "Olivia Prijma"
                  },
                  "phoneNumber": {
                    "description": "Phone number",
                    "type": "string",
                    "example": "+420774839319"
                  },
                  "email": {
                    "description": "Email",
                    "type": "string",
                    "example": "example@mail.com"
                  },
                  "isFavourite": {
                    "description": "Is the contact favourite",
                    "type": "boolean",
                    "example": "true"
                  },
                  "contactType": {
                    "description": "Is the contact favourite",
                    "type": "boolean",
                    "example": "true/false"
                  },
                  "photo": {
                    "description": "Avatar or contact picture",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The successfull",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorisation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "UnauthorizedError"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "message"
                      ],
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Access token expired"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contacts": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "phoneNumber",
          "email",
          "contactType",
          "isFavourite",
          "userId",
          "photo"
        ],
        "properties": {
          "_id": {
            "description": "Contact ID",
            "type": "string",
            "example": "66d9bd6a3b0b0e7e161e1330"
          },
          "name": {
            "description": "Contact name",
            "type": "string",
            "example": "Olivia Prijma"
          },
          "phoneNumber": {
            "description": "Phone number",
            "type": "string",
            "example": "+420774839319"
          },
          "email": {
            "description": "Email",
            "type": "string",
            "example": "example@mail.com"
          },
          "isFavourite": {
            "description": "Value: true/false",
            "type": "string",
            "example": "true"
          },
          "contactType": {
            "description": "The value can be home/work/personal",
            "type": "string",
            "example": "personal"
          },
          "userId": {
            "description": "Id of the parent user",
            "type": "string",
            "example": "66d0e811fb991a743036f812"
          },
          "photo": {
            "description": "You can add photo to a contact",
            "type": "string",
            "example": "https://res.cloudinary.com/do9xmehmv/image/upload/v1725569778/vpril08bt4zyijy4k5bg.jpg"
          }
        }
      }
    }
  }
}